+++ cookbooks/printers/metadata.rb
@@ -1,7 +1,8 @@
-name			 "printers"
-maintainer       "Wesleyan University"
-maintainer_email "software@wesleyan.edu"
-license          "All rights reserved"
-description      "Installs/Configures Campus Printers"
+name			 'printers'
+maintainer       'Wesleyan University'
+maintainer_email 'software@wesleyan.edu'
+license          'All rights reserved'
+description      'Installs/Configures Campus Printers'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.5.5"
+version          '0.6.0'
+depends          'dmg'