+++ cookbooks/printers/recipes/mac.rb
@@ -2,25 +2,12 @@
 # Cookbook Name:: printers
 # Recipe:: mac
 #
-# Copyright 2013, Wesleyan University
+# Copyright 2014, Wesleyan University
 #
 # All rights reserved - Do Not Redistribute
 #
-
-# Install Epson drivers
-dmg_package "EPSON Printer Drivers" do
-  app "EPSONPrinterDrivers"
-  volumes_dir "EPSON Printer Drivers"
-  dmg_name "EPSONPrinterDrivers2.15"
-  source "http://ims-chef.wesleyan.edu/os_x/epson_printer/EPSONPrinterDrivers2.15.dmg"
-  checksum "f9f5658ff11b9933b9f777c15a1d433dfdb1c3131693fb8e0382a5372ba293b6"
-  action :install
-  type "pkg"
-  package_id "com.apple.pkg.EPSONPrinterDrivers"
-  version "2.15.0"
-end
-
-#install Pharos drivers
+#
+# Install Pharos client
 dmg_package "Pharos Popup Printer Client" do
   app "Popup"
   volumes_dir "Popup"
@@ -31,30 +18,6 @@ dmg_package "Pharos Popup Printer Client" do
   type "pkg"
   package_id "com.pharos.popup"
   version "8.4.1"
-end
-
-dmg_package "Epson 3880 Driver" do
-  app "Epson3880"
-  volumes_dir "Epson"
-  dmg_name "Epson"
-  type "mpkg"
-  source "http://ims-chef.wesleyan.edu/os_x/epson_printer/Epson.dmg"
-  checksum "b620b0761a4ad1a5aa6a38124716aede665e6429e8d5b154cb5bb2d467eef247"
-  action :install
-  package_id "com.epson.drivers.3880"
-  version "1.0.0"
-end
-
-dmg_package "Epson 9800 Driver" do
-  app "Epson9800"
-  volumes_dir "Epson"
-  dmg_name "Epson"
-  type "mpkg"
-  source "http://ims-chef.wesleyan.edu/os_x/epson_printer/Epson.dmg"
-  checksum "b620b0761a4ad1a5aa6a38124716aede665e6429e8d5b154cb5bb2d467eef247"
-  action :install
-  package_id "com.epson.drivers.9800"
-  version "1.0.0"
 end
 
 printers "Printers" do